Move KeysignPayload into data module
